@@ -4,74 +4,73 @@
   "scripts": {
     "clean": "rm -rf .next",
     "dev": "NODE_ENV=development nodemon",
-    "build": "NODE_ENV=production npm run clean && next build && tsc --project tsconfig.server.json && cp -r keys .next && cp .env .next && cp static/*.png .next/static",
+    "build": "NODE_ENV=production npm run clean && next build && tsc --project tsconfig.server.json && cp -r keys .next && cp .env .next && cp public/*.png .next/public",
     "start": "node .next/server",
     "deploy": "read -p \"GCP Project ID: \" PROJ_ID; gcloud --project $PROJ_ID app deploy app.yaml --version main --verbosity=debug",
     "stubs": "protoc -I /users/matt/workspace/unitedoperations/uo-discordbot/src/protos --plugin=\"protoc-gen-ts=./node_modules/.bin/protoc-gen-ts\" --js_out=\"import_style=commonjs,binary:./server/lib/grpc\" --ts_out=\"service=true:./server/lib/grpc\" provision.proto"
   },
   "dependencies": {
-    "@google-cloud/connect-datastore": "^3.0.1",
-    "@google-cloud/datastore": "^4.1.3",
-    "@grpc/proto-loader": "^0.5.1",
+    "@google-cloud/connect-datastore": "^4.0.0",
+    "@google-cloud/datastore": "^5.0.4",
+    "@grpc/proto-loader": "^0.5.3",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.4",
     "cors": "^2.8.5",
-    "dotenv": "^6.2.0",
-    "express": "^4.16.4",
-    "express-session": "^1.15.6",
-    "grpc": "^1.19.0",
-    "helmet": "^3.16.0",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "express-session": "^1.17.0",
+    "grpc": "^1.24.2",
+    "helmet": "^3.21.2",
     "morgan": "^1.9.1",
-    "next": "8.0.3",
-    "node-fetch": "^2.3.0",
+    "next": "^9.2.1",
+    "node-fetch": "^2.6.0",
     "node-teamspeak": "^1.0.10",
-    "nodemailer": "^5.1.1",
-    "passport": "^0.4.0",
+    "nodemailer": "^6.4.2",
+    "passport": "^0.4.1",
     "passport-oauth2": "^1.5.0",
-    "react": "16.8.3",
-    "react-dom": "16.8.3",
+    "react": "16.12.0",
+    "react-dom": "16.12.0",
     "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.85.0",
-    "shortid": "^2.2.14",
-    "socket.io": "^2.2.0",
-    "socket.io-client": "^2.2.0",
+    "semantic-ui-react": "^0.88.2",
+    "shortid": "^2.2.15",
+    "socket.io": "^2.3.0",
+    "socket.io-client": "^2.3.0",
     "whatwg-fetch": "^3.0.0",
-    "yaml": "^1.4.0"
+    "yaml": "^1.7.2"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
-    "@types/compression": "0.0.36",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/cors": "^2.8.4",
-    "@types/dotenv": "^6.1.0",
-    "@types/express": "^4.16.1",
-    "@types/express-session": "^1.15.12",
-    "@types/helmet": "0.0.42",
-    "@types/morgan": "^1.7.35",
-    "@types/next": "^8.0.3",
-    "@types/node": "^11.11.5",
-    "@types/node-fetch": "^2.1.7",
-    "@types/nodemailer": "^4.6.6",
-    "@types/passport": "^1.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@types/compression": "1.0.1",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/cors": "^2.8.6",
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.2",
+    "@types/express-session": "^1.15.16",
+    "@types/helmet": "0.0.45",
+    "@types/morgan": "^1.7.37",
+    "@types/next": "^9.0.0",
+    "@types/node": "^13.7.0",
+    "@types/node-fetch": "^2.5.4",
+    "@types/nodemailer": "^6.4.0",
+    "@types/passport": "^1.0.2",
     "@types/passport-oauth2": "^1.4.8",
-    "@types/react": "^16.8.8",
-    "@types/react-dom": "^16.8.3",
+    "@types/react": "^16.9.19",
+    "@types/react-dom": "^16.9.5",
     "@types/shortid": "0.0.29",
-    "@types/socket.io": "^2.1.2",
+    "@types/socket.io": "^2.1.4",
     "@types/socket.io-client": "^1.4.32",
     "@types/whatwg-fetch": "0.0.33",
-    "@types/yaml": "^1.0.2",
+    "@types/yaml": "^1.2.0",
     "@zeit/next-css": "^1.0.1",
-    "@zeit/next-typescript": "^1.1.1",
-    "css-loader": "^1.0.1",
-    "file-loader": "^3.0.1",
-    "google-protobuf": "^3.7.1",
-    "nodemon": "^1.18.10",
+    "css-loader": "^3.4.2",
+    "file-loader": "^5.0.2",
+    "google-protobuf": "^3.11.3",
+    "nodemon": "^2.0.2",
     "protobufjs": "^6.8.8",
-    "ts-node": "^8.0.3",
-    "ts-protoc-gen": "^0.9.0",
-    "typescript": "^3.3.4000",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.29.6"
+    "ts-node": "^8.6.2",
+    "ts-protoc-gen": "^0.12.0",
+    "typescript": "^3.7.5",
+    "url-loader": "^3.0.0",
+    "webpack": "^4.41.5"
   }
 }
