@@ -1,23 +1,18 @@
-import * as React from 'react'
+import React from 'react'
 import Head from 'next/head'
-import 'semantic-ui-css/semantic.min.css'
 
-interface IPageProps {
+interface PageProps {
   title?: string
   children: React.ReactNode
 }
 
-class Page extends React.PureComponent<IPageProps> {
-  render() {
-    return (
-      <div style={{ marginTop: '2em' }}>
-        <Head>
-          <title>{this.props.title || 'United Operations'}</title>
-        </Head>
-        {this.props.children}
-      </div>
-    )
-  }
-}
+const Page: React.FunctionComponent<PageProps> = props => (
+  <div style={{ marginTop: '2em' }}>
+    <Head>
+      <title>{props.title || 'United Operations'}</title>
+    </Head>
+    {props.children}
+  </div>
+)
 
 export default Page
