@@ -1,6 +1,7 @@
-import * as next from 'next'
+import next from 'next'
+import Server from 'next/dist/next-server/server/next-server'
 
-const app: next.Server = next({ dev: process.env.NODE_ENV !== 'production' })
+const app: Server = next({ dev: process.env.NODE_ENV !== 'production' })
 export const nextHandler = app.getRequestHandler()
 
 export default app
